aws/cdk/requirements.txt
aws-cdk-lib==2.136.0
constructs>=10.0.0,<11.0.0
aws-cdk-lib[core]==2.136.0
aws-cdk-lib[aws-ec2]==2.136.0
aws-cdk-lib[aws-ecs]==2.136.0
aws-cdk-lib[aws-rds]==2.136.0
aws-cdk-lib[aws-elasticloadbalancingv2]==2.136.0
aws-cdk-lib[aws-cloudfront]==2.136.0
aws-cdk-lib[aws-s3]==2.136.0
aws-cdk-lib[aws-cloudwatch]==2.136.0
aws-cdk-lib[aws-logs]==2.136.0
aws-cdk-lib[aws-iam]==2.136.0
aws-cdk-lib[aws-secretsmanager]==2.136.0
aws-cdk-lib[aws-sns]==2.136.0
aws-cdk-lib[aws-lambda]==2.136.0
aws-cdk-lib[aws-events]==2.136.0
aws-cdk-lib[aws-events-targets]==2.136.0
aws-cdk-lib[aws-kms]==2.136.0
aws-cdk-lib[aws-ecr]==2.136.0
aws-cdk-lib[aws-autoscaling]==2.136.0
aws-cdk-lib[aws-elasticache]==2.136.0
